Module V1.C_define

module Type : sig ... end

module Value : sig ... end

val import : t -> ?prelude:string -> ?c_flags:string list -> includes:string list -> (string * Type.t) list -> (string * Value.t) list

Import some #define from the given header files. For instance:

        # C.C_define.import c ~includes:"caml/config.h" ["ARCH_SIXTYFOUR", Switch];;
        - (string * Configurator.C_define.Value.t) list = ["ARCH_SIXTYFOUR", Switch true]

val gen_header_file : t -> fname:string -> ?protection_var:string -> (string * Value.t) list -> unit

Generate a C header file containing the following #define. protection_var is used to enclose the file with:

#ifndef BLAH #define BLAH ... #endif

If not specified, it is inferred from the name given to create and the filename.